LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/067-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'script'. Expected ')' to end an argument list.
FAILED (This TC requires JavaScript enabled)

PASS  scheduler: is a script with syntax error marked as "has run"?
